+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 ### Changed
  - Backend changes.
  - Speed up HDR processing.
+### Removed
+ - VAAPI encoders.
 # RVE 2.3.8
 ### Added
  - Open output folder button